:recycle: Refactor select peer and test cases
