--- packages/lisk-p2p/src/errors.ts
@@ -1,27 +1,0 @@
-/*
- * Copyright © 2018 Lisk Foundation
- *
- * See the LICENSE file at the top-level directory of this distribution
- * for licensing information.
- *
- * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
- * no part of this software, including this file, may be copied, modified,
- * propagated, or distributed except according to the terms contained in the
- * LICENSE file.
- *
- * Removal or modification of this copyright notice is prohibited.
- *
- */
-import { VError } from 'verror';
-
-export class NotEnoughPeersError extends VError {
-	private readonly code = 'NOT_ENOUGH_PEERS';
-
-	public constructor(message?: string) {
-		super(message || '');
-		this.name = 'Not Enough Peers Error';
-	}
-	public get Code(): string {
-		return this.code;
-	}
-}
