+++ packages/lisk-p2p/src/p2p.ts
@@ -12,7 +12,6 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-import { NotEnoughPeersError } from './errors';
 import {
 	NetworkStatus,
 	P2PMessagePacket,
@@ -23,119 +22,6 @@ import {
 
 import { Peer } from './peer';
 import { PeerPool } from './peer_pool';
-
-export interface PeerReturnType {
-	readonly options: PeerOptions;
-	readonly peers: ReadonlyArray<Peer>;
-}
-export interface PeerOptions {
-	readonly [key: string]: string | number;
-}
-/* tslint:disable: readonly-keyword*/
-interface Histogram {
-	[key: number]: number;
-}
-interface HistogramValues {
-	height: number;
-	histogram: Histogram;
-	max: number;
-}
-/* tslint:enable: readonly-keyword */
-export const selectPeers = (
-	peers: ReadonlyArray<Peer>,
-	options: PeerOptions,
-	numOfPeers?: number,
-): PeerReturnType => {
-	const filteredPeers = peers.filter(
-		// Remove unreachable peers or heights below last block height
-		(peer: Peer) => peer.Height >= options.blockHeight,
-	);
-
-	if (filteredPeers.length === 0) {
-		const optionsTemp = { ...options, blockHeight: 0 };
-
-		return { options: optionsTemp, peers: [] };
-	}
-
-	// Order peers by descending height
-	const sortedPeers = filteredPeers.sort((a, b) => b.Height - a.Height);
-
-	const aggregation = 2;
-	const returnType: HistogramValues = { height: 0, histogram: {}, max: -1 };
-	const calculatedHistogramValues = sortedPeers.reduce(
-		(
-			histogramValues: HistogramValues = {
-				height: 0,
-				histogram: {},
-				max: -1,
-			},
-			peer: Peer,
-		) => {
-			const val = (peer.Height / aggregation) * aggregation;
-			histogramValues.histogram[val] =
-				(histogramValues.histogram[val] ? histogramValues.histogram[val] : 0) +
-				1;
-			if (histogramValues.histogram[val] > histogramValues.max) {
-				histogramValues.max = histogramValues.histogram[val];
-				histogramValues.height = val;
-			}
-
-			return histogramValues;
-		},
-		returnType,
-	);
-
-	// Perform histogram cut of peers too far from histogram maximum
-	const processedPeers = sortedPeers.filter(
-		peer =>
-			peer &&
-			Math.abs(
-				(calculatedHistogramValues ? calculatedHistogramValues.height : 0) -
-					peer.Height,
-			) <
-				aggregation + 1,
-	);
-
-	// Select n number of peers
-	if (numOfPeers) {
-		if (numOfPeers > processedPeers.length) {
-			throw new NotEnoughPeersError(
-				`Requested no. of peers: '${numOfPeers}' is more than the available no. of good peers: '${
-					processedPeers.length
-				}'`,
-			);
-		}
-		if (numOfPeers === processedPeers.length) {
-			return { options, peers: processedPeers };
-		}
-		if (numOfPeers === 1) {
-			const goodPeer: ReadonlyArray<Peer> = [
-				processedPeers[Math.floor(Math.random() * processedPeers.length)],
-			];
-
-			return { options, peers: goodPeer };
-		}
-		const randomPeersArray = Array(numOfPeers).fill({});
-
-		const selectedPeersList = randomPeersArray.reduce(
-			(peerList: ReadonlyArray<Peer>) => {
-				const peer =
-					processedPeers[Math.floor(Math.random() * processedPeers.length)];
-
-				if (peerList.indexOf(peer) > -1) {
-					return peerList;
-				}
-
-				return [...peerList, peer];
-			},
-			[],
-		);
-
-		return { options, peers: selectedPeersList };
-	}
-
-	return { options, peers: processedPeers };
-};
 
 export class P2P {
 	private readonly peerPool: PeerPool;
